requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
asgiref==3.2.10
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
pep8-naming==0.12.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
telegram==0.0.1
testfixtures==6.18.3
toml==0.10.2
tomli==2.0.0
tornado==6.1
typed-ast==1.5.2
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
zipp==3.7.0
